Renamed getHelp to getDescription
