+++ src/main/java/com/devflask/roboflask/command/moderation/Kick.java
@@ -1,7 +1,7 @@
 package com.devflask.roboflask.command.moderation;
 
-import com.devflask.roboflask.command.Command;
 import com.devflask.roboflask.command.CommandInformation;
+import com.devflask.roboflask.interfaces.Command;
 import com.devflask.roboflask.util.MessageUtil;
 import com.devflask.roboflask.util.Messages;
 import net.dv8tion.jda.api.Permission;
@@ -27,7 +27,7 @@ public class Kick implements Command {
     }
 
     @Override
-    public @NotNull String getHelp() {
+    public @NotNull String getDescription() {
         return "Kicks the member specified - !kick <@Person> [reason]";
     }
