+++ src/main/java/com/devflask/roboflask/command/util/Help.java
@@ -1,7 +1,7 @@
 package com.devflask.roboflask.command.util;
 
-import com.devflask.roboflask.command.Command;
 import com.devflask.roboflask.command.CommandInformation;
+import com.devflask.roboflask.interfaces.Command;
 import com.devflask.roboflask.util.MessageUtil;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.events.message.priv.PrivateMessageReceivedEvent;
@@ -29,7 +29,7 @@ public class Help implements Command {
     }
 
     @Override
-    public @NotNull String getHelp() {
+    public @NotNull String getDescription() {
         return "Retrieves help from the bot.";
     }
 
@@ -41,11 +41,11 @@ public class Help implements Command {
         StringBuilder commands = new StringBuilder();
         HashMap<String, String> commandsMap = new HashMap<>();
         for (Command c : getManager().getCommands()){
-            System.out.println(c.getName()+" - "+c.getHelp());
+            System.out.println(c.getName() + " - " + c.getDescription());
             commands
                     .append(c.getName())
                     .append(" - ")
-                    .append(c.getHelp())
+                    .append(c.getDescription())
                     .append("\n")
             ;
         }
